STYLE: add parameter to not write the interpolation date file

+++ iota2/Sensors/Sentinel_2.py
@@ -363,7 +363,7 @@ class Sentinel_2(Sensor):
 
         return superimp, app_dep
 
-    def write_interpolation_dates_file(self):
+    def write_interpolation_dates_file(self, write=True):
         """
         TODO : mv to base-class
         """
@@ -382,7 +382,7 @@ class Sentinel_2(Sensor):
             date_interp_max = self.cfg_IOTA2.getParam("Sentinel_2", "endDate")
 
         dates = [str(date).replace("-","") for date in dateInterval(date_interp_min, date_interp_max, self.temporal_res)]
-        if not os.path.exists(interp_date_file):
+        if not os.path.exists(interp_date_file) and write:
             with open(interp_date_file, "w") as interpolation_date_file:
                 interpolation_date_file.write("\n".join(dates))
         return interp_date_file, dates
